Simplify ANSI percentile override via perc_adj_factor
Add perc_adj_factor param to DialectDefinition.render_percentile
(default 1.0, no-op). AnsiDialectDef overrides the default to 100.0
so approx_percentile receives a fraction, replacing the manual
AggCol/PercentileConfig copy. Revert PercentileConfig.percentile
back to int.

+++ src/casalib/sql_connection/dialects/helpers.py
@@ -54,7 +54,7 @@ class AggOperationEnum(Enum):
 class PercentileConfig:
     """Configuration for a percentile aggregation."""
 
-    percentile: float
+    percentile: int
     ignore_values: List[float] = field(
         default_factory=list
     )
